Separacion de modelos de releamiento
